backend/requirements.txt
# Core
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7

# Authentication
PyJWT==2.8.0
passlib==1.7.4

# Email
Flask-Mail==0.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0
